use floats
git-svn-id: http://svn.macosforge.org/repository/ruby/MacRuby/trunk@401 23306eb0-4c56-4727-a40e-e92c0eb68959

+++ sample-macruby/HotCocoa/calculator.rb
@@ -84,7 +84,7 @@ class Calculator
     end
   
     def sqrt
-      value.text = Math.sqrt(value.to_i).to_s
+      value.text = Math.sqrt(value.to_f).to_s
     end
   
     def clear
